Add files via upload
I updated the agent's perception of the gradients to more closely match what is known about how organisms interperet sensory cues. Whether or not this solves the original problem that we had of the behaviour of the std in 'angle-change', or if it changes any of those calculations we should discuss. I also added the time it takes each bacteria to reach the source for the first time, and also a measure of how long they stay near the source."